refactor: Store - common, invitation 경로 수정
