+++ src/components/user/UserInvitationListEdit.tsx
@@ -8,8 +8,8 @@ import { UserInvitationListEditProps } from '@/types/invitation/edit';
 import { getInvitationDeleteList } from '@/pages/api/invitation/editRequests';
 import theme from '@/styles/theme';
 import Modal from '@/components/common/Modal';
-import useModalStore from '@/stores/useModalStore';
-import useBottomSheetStore from '@/stores/useBottomSheetStore';
+import useModalStore from '@/stores/common/useModalStore';
+import useBottomSheetStore from '@/stores/common/useBottomSheetStore';
 
 function UserInvitationListEdit({ id }: UserInvitationListEditProps) {
   const router = useRouter();
